# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Bruma API
  version: 1.0.0
  description: |
    MVP HTTP API (tenants, domains, templates, dispatch, webhooks, logs). Authenticate with `X-API-Key`
    (admin key or per-tenant key). Webhook endpoint uses HMAC signature headers instead of API keys.

    **Swagger UI:** `GET /docs` when `OPENAPI_DOCS_ENABLED` is true or outside production (see deployment config).
servers:
  - url: http://localhost:3000
    description: Local
...
paths:
...
  /v1/webhooks/provider:
    post:
      tags: [Webhooks]
      summary: Provider delivery webhook (HMAC-signed raw body)
      description: |
        Send raw JSON body. Signature: `X-Webhook-Signature` or `X-Bruma-Webhook-Signature` (HMAC-SHA256 hex of body using `WEBHOOK_SECRET`).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WebhookPayload" }
      responses:
        "204": { description: Applied }
        "400": { description: Invalid JSON or validation }
        "401": { description: Bad signature }
        "404": { description: Unknown provider_message_id }
        "403": { description: Tenant inactive — no update applied }
        "503": { description: WEBHOOK_SECRET not set }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  schemas:
...
    WebhookPayload:
      type: object
      required: [provider_message_id, event]
      properties:
        provider_message_id: { type: string }
        event:
          type: string
          enum: [delivered, bounced, spam_complaint, failed, sent]
        occurred_at: { type: string }
        recipient: { type: string, format: email }
...
security:
  - ApiKeyAuth: []